added socket events

+++ src/controllers/twitterController.js
@@ -1,6 +1,6 @@
-import { catchAsync } from "../utils"
-import { db, storage } from "../configs"
-import { userT, followT, tweetT, likeT, notificationT, tweetSchema, arrayAggOrder, arrayAgg, alias, commentTweetT } from "../drizzle/schema"
+import { catchAsync, roomKey } from "../utils"
+import { db, redis, storage } from "../configs"
+import { userT, followT, tweetT, likeT, notificationT, tweetSchema, alias } from "../drizzle/schema"
 import { eq, ne, ilike, and, sql, desc } from "drizzle-orm"
 import { mediaInput } from "../validators"
 import { io } from "../socket"
@@ -10,23 +10,23 @@ export const getFeed = catchAsync(async (req, res) => {
   const { id } = req.user
 
   const parentTweetT = alias(tweetT, "ParentTweet")
-  const parentTweetUserT = alias(tweetT, "ParentTweetUser")
+  const parentTweetUserT = alias(userT, "ParentTweetUser")
+  const parentTweetLikeT = alias(likeT, "ParentTweetLike")
+
   const feed = await db
     .select({
-      ...tweetT,
-      user: userT,
-      isLiked: likeT,
-      parentTweet: {
-        ...parentTweetT,
-        user: parentTweetUserT,
-      },
+      tweet: { ...tweetT, user: userT },
+      isTweetLiked: likeT,
+      parentTweet: { ...parentTweetT, user: parentTweetUserT },
+      isParentTweetLiked: parentTweetLikeT,
     })
     .from(followT)
     .innerJoin(tweetT, and(eq(id, followT.from), eq(followT.to, tweetT.userId)))
     .leftJoin(userT, eq(tweetT.userId, userT.id))
+    .leftJoin(likeT, and(eq(tweetT.id, likeT.tweetId), eq(likeT.userId, id)))
     .leftJoin(parentTweetT, eq(tweetT.parentTweetId, parentTweetT.id))
     .leftJoin(parentTweetUserT, eq(parentTweetT.userId, parentTweetUserT.id))
-    .leftJoin(likeT, and(eq(tweetT.id, likeT.tweetId), eq(likeT.userId, id)))
+    .leftJoin(parentTweetLikeT, and(eq(parentTweetT.id, parentTweetLikeT.tweetId), eq(parentTweetLikeT.userId, id)))
     .orderBy(desc(tweetT.createdAt))
 
   res.json({ data: feed })
@@ -82,8 +82,8 @@ export const getPublicProfile = catchAsync(async (req, res) => {
   const userQ = db
     .select({
       ...userT,
-      followers: sql`(select count(*)-1 from ${followT} where ${followT.to} = ${userId})`,
-      following: sql`(select count(*)-1 from ${followT} where ${followT.from} = ${userId})`,
+      followers: sql`(select count(*) from ${followT} where ${followT.to} = ${userId})`,
+      following: sql`(select count(*) from ${followT} where ${followT.from} = ${userId})`,
       isFollowed: sql`(select count(*) from ${followT} where ${followT.from} = ${id} and ${followT.to} = ${userId})`,
     })
     .from(userT)
@@ -141,26 +141,32 @@ export const search = catchAsync(async (req, res) => {
 export const createTweet = catchAsync(async (req, res) => {
   const { id } = req.user
   const { parentTweetId } = req.body
-
+  // check and upload media if exist
   if (req?.files?.media) {
     const [media, mediaType] = mediaInput(req.files.media)
     const mediaURL = await storage.upload(uuid(), media.data)
-
     req.body.mediaType = mediaType
     req.body.mediaURL = mediaURL
   }
-
+  // check and create tweet
   const data = tweetSchema.parse({ ...req.body, userId: id })
   const [tweet] = await db.insert(tweetT).values(data).returning()
-
-  if (parentTweetId)
+  // create redis cache
+  await redis.hSet(roomKey("tweet", tweet.id), { likes: 0, comments: 0 })
+  // check if its a reply tweet
+  if (parentTweetId) {
+    // update and emit comment count
+    const key = roomKey("tweet", parentTweetId)
+    const comments = await redis.hIncrBy(key, "comments", 1)
+    io.in(key).emit(roomKey(key, "comments"), comments)
+    // get parent tweet
     db.select()
       .from(tweetT)
       .where(eq(parentTweetId, tweetT.id))
       .then(([tweet]) => {
-        // check if self reply
+        // don't send notification if self reply
         if (tweet.userId === id) return
-
+        // create and send notification
         db.insert(notificationT)
           .values({
             type: "comment",
@@ -168,8 +174,9 @@ export const createTweet = catchAsync(async (req, res) => {
             from: id,
             tweetId: tweet.id,
           })
-          .then(console.log)
+          .then(() => io.in(roomKey("user", tweet.userId)).emit("new_notification"))
       })
+  }
 
   res.json({ data: { ...tweet, user: req.user } })
 })
@@ -177,16 +184,20 @@ export const createTweet = catchAsync(async (req, res) => {
 export const like = catchAsync(async (req, res) => {
   const { id } = req.user
   const { tweetId } = req.params
-
+  // create like
   await db.insert(likeT).values({ userId: id, tweetId })
-
+  // update and emit like count
+  const key = roomKey("tweet", tweetId)
+  const likes = await redis.hIncrBy(key, "likes", 1)
+  io.in(key).emit(roomKey(key, "likes"), likes)
+  // get tweet
   db.select()
     .from(tweetT)
     .where(eq(tweetId, tweetT.id))
     .then(([tweet]) => {
-      // check if self like
+      // don't send notification if self like
       if (tweet.userId === id) return
-
+      // create and send notification
       db.insert(notificationT)
         .values({
           type: "like",
@@ -194,7 +205,7 @@ export const like = catchAsync(async (req, res) => {
           from: id,
           tweetId: tweet.id,
         })
-        .then(console.log)
+        .then(() => io.in(roomKey("user", tweet.userId)).emit("new_notification"))
     })
 
   res.end()
@@ -203,8 +214,18 @@ export const like = catchAsync(async (req, res) => {
 export const unlike = catchAsync(async (req, res) => {
   const { id } = req.user
   const { tweetId } = req.params
-
-  await db.delete(likeT).where(and(eq(likeT.userId, id), eq(likeT.tweetId, tweetId)))
+  // delete unlike
+  const [unlike] = await db
+    .delete(likeT)
+    .where(and(eq(likeT.userId, id), eq(likeT.tweetId, tweetId)))
+    .returning()
+  // if unlike successful
+  if (unlike) {
+    // update and emit like count
+    const key = roomKey("tweet", tweetId)
+    const likes = await redis.hIncrBy(key, "likes", -1)
+    io.in(key).emit(roomKey(key, "likes"), likes)
+  }
 
   res.end()
 })
@@ -212,16 +233,18 @@ export const unlike = catchAsync(async (req, res) => {
 export const follow = catchAsync(async (req, res) => {
   const { id } = req.user
   const { userId } = req.params
-
+  // check for self follow
+  if (id === userId) return res.status(400).end()
+  // create follow
   await db.insert(followT).values({ to: userId, from: id })
-
+  // create and send notification
   db.insert(notificationT)
     .values({
       type: "follow",
       to: userId,
       from: id,
     })
-    .then(console.log)
+    .then(() => io.in(roomKey("user", userId)).emit("new_notification"))
 
   res.end()
 })
@@ -229,7 +252,9 @@ export const follow = catchAsync(async (req, res) => {
 export const unfollow = catchAsync(async (req, res) => {
   const { id } = req.user
   const { userId } = req.params
-
+  // check for self follow
+  if (id === userId) return res.status(400).end()
+  // delete follow
   await db.delete(followT).where(and(eq(followT.to, userId), eq(followT.from, id)))
 
   res.end()
